Drops pattern grammar and multi-lang support.
As discussed in the paper pattern is straightforward to lex, so the
grammar adds unnecesary and brittle complexity. For example, strings
were incorrectly parsed due to the fragmented lexing.

Multi-language support was originally introduced to support custom
language and has shown to make the Document API cumbersome. Therefore,
we also drop it and inline the unifying Tree API for simplicity.